+++ projects/ngx-smooth-scroll/CHANGELOG.md
@@ -1,3 +1,46 @@
+#### 0.3.0-beta Add methods & properties to each directive
+- `NgxSmoothScrollWheelDirective.containerEl`: HTMLElement, container element.
+- `NgxSmoothScrollWheelDirective.children`: HTMLElement[], pure array of child elements.
+
+- `NgxSmoothScrollWheelDirective.scrollTo`: (destination, options) => Observable<{ x: number; y: number; }>
+
+    Scroll to given destination.
+    - `destination`: object, `{ x: number; y: number; }`, `required`
+    - `options`: NgxSmoothScrollOption, see more [here](https://github.com/Seemspyo/ngx-smooth-scroll#scroll-option)
+
+- `NgxSmoothScrollWheelDirective.scrollToChild`: (childEl, options) => Observable<{ x: number; y: number; }>
+
+    Scroll to  given child element
+    - `childEl`: HTMLElement, `required`
+    - `options`: NgxSmoothScrollOption, see more [here](https://github.com/Seemspyo/ngx-smooth-scroll#scroll-option)
+
+- `NgxSmoothScrollWheelDirective.scrollToIndex`: (index, options) => Observable<{ x: number; y: number; }>
+
+    Scroll to given child index. childSelector must be set.
+    - `index`: number, index of child element, `required`
+    - `options`: NgxSmoothScrollOption, see more [here](https://github.com/Seemspyo/ngx-smooth-scroll#scroll-option)
+
+- `NgxSmoothScrollKeyboardDirective.containerEl`: HTMLElement, container element.
+- `NgxSmoothScrollKeyboardDirective.children`: HTMLElement[], pure array of child elements.
+
+- `NgxSmoothScrollKeyboardDirective.scrollTo`: (destination, options) => Observable<{ x: number; y: number; }>
+
+    Scroll to given destination.
+    - `destination`: object, `{ x: number; y: number; }`, `required`
+    - `options`: NgxSmoothScrollOption, see more [here](https://github.com/Seemspyo/ngx-smooth-scroll#scroll-option)
+
+- `NgxSmoothScrollKeyboardDirective.scrollToChild`: (childEl, options) => Observable<{ x: number; y: number; }>
+
+    Scroll to  given child element
+    - `childEl`: HTMLElement, `required`
+    - `options`: NgxSmoothScrollOption, see more [here](https://github.com/Seemspyo/ngx-smooth-scroll#scroll-option)
+
+- `NgxSmoothScrollKeyboardDirective.scrollToIndex`: (index, options) => Observable<{ x: number; y: number; }>
+
+    Scroll to given child index. childSelector must be set.
+    - `index`: number, index of child element, `required`
+    - `options`: NgxSmoothScrollOption, see more [here](https://github.com/Seemspyo/ngx-smooth-scroll#scroll-option)
+
 #### 0.2.3-beta fix interrupt method
 #### 0.2.2-beta hot fix for child selector
 #### 0.2.1-beta Add directives to exports